requirements.txt
pandas
plotly
streamlit
pillow
xmltodict
paramiko
colorutils
PyYAML
